requirements.txt
hydra-core==1.3.2
invoke==2.2.0
loguru==0.7.3
matplotlib==3.10.0
python-dotenv==1.0.1
scikit_learn==1.6.0
torch==2.5.1
torchvision==0.20.1
wandb==0.19.2
fastapi==0.115.6
Pillow==11.1.0
pytest==8.3.4
uvicorn==0.23.2
pydantic==2.10.5
python-multipart==0.0.20
httpx==0.28.1
numpy==1.26.3
google-cloud-storage==2.19.0